@@ -1,8 +1,6 @@
 {
   "name": "snapshot",
   "version": "0.1.3",
-  "repository": "balancer-labs/snapshot",
-  "license": "MIT",
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
@@ -43,7 +41,6 @@
     "jsonexport": "^3.0.1",
     "lodash": "^4.17.15",
     "numeral": "^2.0.4",
-    "portal-vue": "^2.1.7",
     "pretty-ms": "^7.0.0",
     "primer-support": "^5.0.0",
     "remarkable": "^2.0.1",
@@ -52,17 +49,16 @@
     "stylus": "^0.54.8",
     "stylus-loader": "^3.0.2",
     "typescript": "~3.9.3",
-    "vue": "^2.6.11",
-    "vue-autofocus-directive": "^1.0.4",
+    "vue": "^3.0.0-beta.1",
     "vue-cli-plugin-webpack-bundle-analyzer": "^2.0.0",
-    "vue-clipboard2": "^0.3.1",
-    "vue-i18n": "^8.18.1",
-    "vue-infinite-scroll": "^2.0.2",
+    "vue-i18n": "^9.0.0-rc.1",
     "vue-jazzicon": "^0.1.3",
-    "vue-router": "^3.2.0",
-    "vue-textarea-autosize": "^1.1.1",
-    "vuedraggable": "^2.24.1",
-    "vuex": "^3.4.0",
+    "vue-router": "^4.0.0-alpha.6",
+    "vue3-clipboard": "^1.0.0",
+    "vue3-infinite-scroll": "^3.0.4",
+    "vue3-jazzicon": "^0.1.1",
+    "vuedraggable": "^4.0.1",
+    "vuex": "^4.0.0-alpha.1",
     "walletlink": "^2.0.2"
   },
   "devDependencies": {
@@ -71,13 +67,16 @@
     "@typescript-eslint/parser": "^2.33.0",
     "@vue/cli-plugin-router": "^4.4.0",
     "@vue/cli-plugin-vuex": "^4.4.0",
+    "@vue/compiler-sfc": "^3.0.0-beta.1",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^5.0.2",
     "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-vue": "^6.2.2",
+    "eslint-plugin-vue": "^7.0.0-alpha.0",
     "prettier": "^1.19.1",
     "sass": "^1.29.0",
     "sass-loader": "^8.0.2",
-    "vue-template-compiler": "^2.6.11"
-  }
+    "vue-cli-plugin-vue-next": "~0.1.4"
+  },
+  "license": "MIT",
+  "repository": "balancer-labs/snapshot"
 }
