@@ -7,7 +7,7 @@
         <router-view :key="$route.path" class="flex-auto" />
       </div>
     </div>
-    <portal-target name="modal" multiple />
+    <div id="modal" />
     <Notifications />
   </div>
 </template>
