@@ -10,20 +10,21 @@
 <script>
 export default {
   props: {
-    value: Number
+    modelValue: Number
   },
+  emits: ['update:modelValue'],
   data() {
     return {
       input: ''
     };
   },
   created() {
-    if (this.value) this.input = this.value.toString();
+    if (this.modelValue) this.input = this.modelValue.toString();
   },
   methods: {
     handleInput() {
-      if (!this.input) return this.$emit('input', undefined);
-      this.$emit('input', parseFloat(this.input));
+      if (!this.input) return this.$emit('update:modelValue', undefined);
+      this.$emit('update:modelValue', parseFloat(this.input));
     }
   }
 };
