@@ -1,6 +1,6 @@
 <template>
   <UiModal :open="open" @close="$emit('close')">
-    <template slot="header">
+    <template v-slot:header>
       <h3 v-if="!web3.account || step === 'connect'">Connect wallet</h3>
       <h3 v-else>Account</h3>
     </template>
@@ -41,7 +41,7 @@
       </div>
     </div>
     <div v-else>
-      <div v-if="$auth.isAuthenticated" class="m-4">
+      <div v-if="$auth.isAuthenticated.value" class="m-4">
         <a
           :href="_explorer(web3.network.key, web3.account)"
           target="_blank"
@@ -105,6 +105,7 @@ import { getInjected } from '@snapshot-labs/lock/src/utils';
 
 export default {
   props: ['open'],
+  emits: ['login', 'close'],
   data() {
     return {
       step: null,
