@@ -1,6 +1,6 @@
 <template>
   <UiModal :open="open" v-if="open" @close="$emit('close')" class="d-flex">
-    <template slot="header">
+    <template v-slot:header>
       <h3>Confirm vote</h3>
     </template>
     <div class="d-flex flex-column flex-auto">
@@ -23,7 +23,7 @@
             target="_blank"
             class="float-right"
           >
-            {{ $n(proposal.msg.payload.snapshot) }}
+            {{ proposal.msg.payload.snapshot }}
             <Icon name="external-link" class="ml-1" />
           </a>
         </div>
@@ -43,7 +43,7 @@
         </div>
       </div>
     </div>
-    <template slot="footer">
+    <template v-slot:footer>
       <div class="col-6 float-left pr-2">
         <UiButton @click="$emit('close')" type="button" class="width-full">
           Cancel
@@ -78,6 +78,7 @@ export default {
     'totalScore',
     'scores'
   ],
+  emits: ['reload', 'close'],
   data() {
     return {
       loading: false
