@@ -1,6 +1,6 @@
 <template>
   <UiModal :open="open" @close="$emit('close')">
-    <template slot="header">
+    <template v-slot:header>
       <h3>Networks</h3>
     </template>
     <div class="mt-4 mx-0 mx-md-4">
@@ -21,6 +21,7 @@ import { filterNetworks } from '@/helpers/utils';
 
 export default {
   props: ['open'],
+  emits: ['update:modelValue', 'close'],
   computed: {
     networks() {
       return filterNetworks(networks, this.app.spaces, '');
@@ -28,7 +29,7 @@ export default {
   },
   methods: {
     select(key) {
-      this.$emit('input', key);
+      this.$emit('update:modelValue', key);
       this.$emit('close');
     }
   }
