@@ -1,6 +1,6 @@
 <template>
   <UiModal :open="open" @close="$emit('close')">
-    <template slot="header">
+    <template v-slot:header>
       <h3>Plugins</h3>
     </template>
     <div class="m-4 mt-4" v-if="selected === false">
@@ -27,7 +27,7 @@
         </UiButton>
       </div>
     </div>
-    <template v-if="selected === false" slot="footer">
+    <template v-if="selected === false" v-slot:footer>
       <div class="col-6 float-left pr-2">
         <UiButton @click="$emit('close')" type="button" class="width-full">
           Cancel
@@ -35,7 +35,7 @@
       </div>
       <div class="col-6 float-left pl-2">
         <UiButton
-          @click="[$emit('input', form), $emit('close')]"
+          @click="[$emit('update:modelValue', form), $emit('close')]"
           type="submit"
           class="width-full button--submit"
         >
@@ -43,16 +43,14 @@
         </UiButton>
       </div>
     </template>
-    <div v-else class="m-4 p-4 border rounded-2 text-white">
+    <div v-if="selected !== false" class="m-4 p-4 border rounded-2 text-white">
       <PluginAragonConfig
-        :value="form.aragon"
         :proposal="proposal"
         v-model="form.aragon"
         @close="selected = false"
         v-if="selected === 'aragon'"
       />
       <PluginGnosisConfig
-        :value="form.gnosis"
         :proposal="proposal"
         :network="space.network"
         v-model="form.gnosis"
@@ -68,7 +66,8 @@ import plugins from '@snapshot-labs/snapshot.js/src/plugins';
 import { clone } from '@/helpers/utils';
 
 export default {
-  props: ['open', 'value', 'space', 'proposal'],
+  props: ['open', 'modelValue', 'space', 'proposal'],
+  emits: ['close', 'update:modelValue'],
   data() {
     return {
       plugins: [],
@@ -78,7 +77,7 @@ export default {
   },
   watch: {
     open() {
-      if (this.value && this.open) this.form = clone(this.value);
+      if (this.modelValue && this.open) this.form = clone(this.modelValue);
       this.selected = false;
     }
   },
