@@ -2,7 +2,7 @@
   <div class="d-flex flex-items-center">
     <Icon name="search" size="22" class="mb-1 mr-2 text-gray" />
     <input
-      :value="value"
+      :value="modelValue"
       :placeholder="placeholder"
       @input="handleInput"
       type="text"
@@ -11,23 +11,24 @@
       class="border-0 input flex-auto width-full"
     />
     <a @click="clearInput">
-      <Icon v-if="value" name="close" size="12" class="mb-1" />
+      <Icon v-if="modelValue" name="close" size="12" class="mb-1" />
     </a>
   </div>
 </template>
 
 <script>
 export default {
-  props: ['value', 'placeholder'],
+  props: ['modelValue', 'placeholder'],
+  emits: ['update:modelValue'],
   methods: {
     handleInput(e) {
       const input = e.target.value;
       this.$router.push({ query: input ? { q: input } : {} });
-      this.$emit('input', input);
+      this.$emit('update:modelValue', input);
     },
     clearInput() {
       this.$router.push({});
-      this.$emit('input', '');
+      this.$emit('update:modelValue', '');
     }
   }
 };
