@@ -1,5 +1,4 @@
-import Vue from 'vue';
-import VueRouter, { RouteConfig } from 'vue-router';
+import { createRouter, createWebHashHistory } from 'vue-router';
 import domains from '@snapshot-labs/snapshot-spaces/spaces/domains.json';
 import aliases from '@snapshot-labs/snapshot-spaces/spaces/aliases.json';
 import Home from '@/views/Home.vue';
@@ -12,7 +11,6 @@ import Explore from '@/views/Explore.vue';
 import Delegate from '@/views/Delegate.vue';
 import Timeline from '@/views/Timeline.vue';
 
-Vue.use(VueRouter);
 const domainName = window.location.hostname;
 
 const beforeEnter = (to: any, from, next) => {
@@ -23,7 +21,7 @@ const beforeEnter = (to: any, from, next) => {
   next();
 };
 
-const routes: Array<RouteConfig> = [
+const routes: any[] = [
   { path: '/setup', name: 'setup', component: Setup },
   { path: '/:key/settings/:from?', name: 'settings', component: Settings },
   { path: '/networks', name: 'networks', component: Explore },
@@ -64,10 +62,10 @@ const routes: Array<RouteConfig> = [
   { path: '/*', name: 'error-404', beforeEnter: (to, from, next) => next('/') }
 ];
 
-const router = new VueRouter({
-  mode: 'hash',
+const router = createRouter({
+  history: createWebHashHistory(),
   routes,
-  scrollBehavior() {
+  scrollBehavior(): {} {
     return { x: 0, y: 0 };
   }
 });
