@@ -1,10 +1,9 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
+import { createStore } from 'vuex';
 import modules from '@/store/modules';
 
-Vue.use(Vuex);
-
-const store = new Vuex.Store({
+// TODO: fix this typing error
+// @ts-ignore
+const store = createStore({
   modules,
   strict: process.env.NODE_ENV !== 'production'
 });
