@@ -1,6 +1,5 @@
-import Vue from 'vue';
 import { Web3Provider } from '@ethersproject/providers';
-import { getInstance } from '@snapshot-labs/lock/plugins/vue';
+import { getInstance } from '@snapshot-labs/lock/plugins/vue3';
 import networks from '@snapshot-labs/snapshot.js/src/networks.json';
 import store from '@/store';
 import { formatUnits } from '@ethersproject/units';
@@ -34,12 +33,12 @@ const mutations = {
         unknown: true
       };
     }
-    Vue.set(_state, 'network', networks[chainId]);
+    _state.network = networks[chainId];
     console.debug('HANDLE_CHAIN_CHANGED', chainId);
   },
   WEB3_SET(_state, payload) {
     Object.keys(payload).forEach(key => {
-      Vue.set(_state, key, payload[key]);
+      _state[key] = payload[key];
     });
   }
 };
@@ -49,25 +48,26 @@ const actions = {
     auth = getInstance();
     commit('SET', { authLoading: true });
     await auth.login(connector);
-    if (auth.provider) {
-      auth.web3 = new Web3Provider(auth.provider);
+    if (auth.provider.value) {
+      auth.web3 = new Web3Provider(auth.provider.value);
       await dispatch('loadProvider');
     }
     commit('SET', { authLoading: false });
   },
   logout: async ({ commit }) => {
-    Vue.prototype.$auth.logout();
+    auth = getInstance();
+    auth.logout();
     commit('WEB3_SET', { account: null, profile: null });
   },
   loadProvider: async ({ commit, dispatch }) => {
     try {
-      if (auth.provider.removeAllListeners && !auth.provider.isTorus)
-        auth.provider.removeAllListeners();
-      if (auth.provider.on) {
-        auth.provider.on('chainChanged', async chainId => {
+      if (auth.provider.value.removeAllListeners)
+        auth.provider.value.removeAllListeners();
+      if (auth.provider.value.on) {
+        auth.provider.value.on('chainChanged', async chainId => {
           commit('HANDLE_CHAIN_CHANGED', parseInt(formatUnits(chainId, 0)));
         });
-        auth.provider.on('accountsChanged', async accounts => {
+        auth.provider.value.on('accountsChanged', async accounts => {
           if (accounts.length !== 0) {
             commit('WEB3_SET', { account: accounts[0] });
             await dispatch('loadProvider');
@@ -75,7 +75,7 @@ const actions = {
         });
         // auth.provider.on('disconnect', async () => {});
       }
-      console.log('Provider', auth.provider);
+      console.log('Provider', auth.provider.value);
       let network, accounts;
       try {
         [network, accounts] = await Promise.all([
